NF: split AudioRecorderTest
With part that requireAndroid and part that don't
